requirements.txt
yfinance>=0.2.36
finnhub-python>=2.4.19
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.0
pandas>=2.2.0
